Update red text in dark mode for better contrast
This CSS change adjusts the red color in dark mode to something with a bit more contrast.

Running HTML CodeSniffer against the site, it yielded an error on the red text. With this change, I don't get that error in dark mode any more. I'm honestly not sure how to easily test in non-dark mode to make sure I didn't break that.